update doc & clean up structure
