+++ docs/refs.bib
@@ -84,3 +84,14 @@
     title = {Towards efficient multiobjective optimization: Multiobjective statistical criterions},
     year = {2012}
 }
+
+@article{yang2019efficient,
+  title={Efficient computation of expected hypervolume improvement using box decomposition algorithms},
+  author={Yang, Kaifeng and Emmerich, Michael and Deutz, Andr{\'e} and B{\"a}ck, Thomas},
+  journal={Journal of Global Optimization},
+  volume={75},
+  number={1},
+  pages={3--34},
+  year={2019},
+  publisher={Springer}
+}
