+++ tests/unit/acquisition/test_function.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 from __future__ import annotations
 
+import itertools
 import math
 import unittest.mock
 from collections.abc import Mapping
@@ -39,22 +40,27 @@ from trieste.acquisition.function import (
     BatchMonteCarloExpectedImprovement,
     BatchReparametrizationSampler,
     ExpectedConstrainedImprovement,
+    ExpectedHypervolumeImprovement,
     ExpectedImprovement,
     IndependentReparametrizationSampler,
     MinValueEntropySearch,
     NegativeLowerConfidenceBound,
     ProbabilityOfFeasibility,
     SingleModelAcquisitionBuilder,
+    expected_hv_improvement,
     expected_improvement,
+    get_reference_point,
     lower_confidence_bound,
     min_value_entropy_search,
     probability_of_feasibility,
 )
 from trieste.data import Dataset
 from trieste.models import ProbabilisticModel
+from trieste.models.model_interfaces import ModelStack
 from trieste.space import Box
 from trieste.type import TensorType
 from trieste.utils.objectives import BRANIN_MINIMUM, branin
+from trieste.utils.pareto import Pareto
 
 
 class _ArbitrarySingleBuilder(SingleModelAcquisitionBuilder):
@@ -459,6 +465,145 @@ def test_expected_constrained_improvement_min_feasibility_probability_bound_is_i
     ei = ExpectedImprovement().using("foo").prepare_acquisition_function(data, models_)
     x = tf.constant([[1.5]])
     npt.assert_allclose(eci(x), ei(x) * pof(x))
+
+
+def _linear_mean_gaussian_process(
+    kernel_amplitude: float | TensorType | None = None,
+) -> GaussianProcess:
+    return GaussianProcess(
+        [lambda x: tf.reduce_sum(x, axis=-1, keepdims=True)],
+        [tfp.math.psd_kernels.ExponentiatedQuadratic(kernel_amplitude)],
+    )
+
+
+_mo_test_models = (
+    QuadraticMeanAndRBFKernel,
+    _linear_mean_gaussian_process,
+    QuadraticMeanAndRBFKernel,
+)
+
+
+def test_ehvi_builder_raises_for_empty_data() -> None:
+    num_obj = 3
+    dataset = Dataset(tf.zeros([0, 2]), tf.zeros([0, 1]))
+    model = ModelStack(*[(QuadraticMeanAndRBFKernel(), 1) for _ in range(num_obj)])
+
+    with pytest.raises(TF_DEBUGGING_ERROR_TYPES):
+        ExpectedHypervolumeImprovement().prepare_acquisition_function(dataset, model)
+
+
+def test_ehvi_builder_builds_expected_hv_improvement_using_pareto_from_model() -> None:
+    num_obj = 2
+    train_x = tf.constant([[-2.0], [-1.5], [-1.0], [0.0], [0.5], [1.0], [1.5], [2.0]])
+    dataset = Dataset(
+        train_x,
+        tf.tile(
+            tf.constant([[4.1], [0.9], [1.2], [0.1], [-8.8], [1.1], [2.1], [3.9]]), [1, num_obj]
+        ),
+    )
+
+    model = ModelStack(*[(_mo_test_models[_](), 1) for _ in range(num_obj)])
+    acq_fn = ExpectedHypervolumeImprovement().prepare_acquisition_function(dataset, model)
+
+    model_pred_observation = model.predict(train_x)[0]
+    _prt = Pareto(model_pred_observation)
+    xs = tf.linspace([[-10.0]], [[10.0]], 100)
+    expected = expected_hv_improvement(model, _prt, get_reference_point(_prt.front))(xs)
+    npt.assert_allclose(acq_fn(xs), expected)
+
+
+@pytest.mark.parametrize("at", [tf.constant([[0.0], [1.0]]), tf.constant([[[0.0], [1.0]]])])
+def test_ehvi_raises_for_invalid_batch_size(at: TensorType) -> None:
+    num_obj = 2
+    train_x = tf.constant([[-2.0], [-1.5], [-1.0], [0.0], [0.5], [1.0], [1.5], [2.0]])
+
+    model = ModelStack(*[(_mo_test_models[_](), 1) for _ in range(num_obj)])
+    model_pred_observation = model.predict(train_x)[0]
+    _prt = Pareto(model_pred_observation)
+    ehvi = expected_hv_improvement(model, _prt, get_reference_point(_prt.front))
+
+    with pytest.raises(TF_DEBUGGING_ERROR_TYPES):
+        ehvi(at)
+
+
+@random_seed
+@pytest.mark.parametrize(
+    "input_dim, num_samples_per_point, existing_observations, obj_num, variance_scale ,rtol, atol",
+    [
+        (
+            1,
+            50_000,
+            tf.constant([[0.3, 0.2], [0.2, 0.22], [0.1, 0.25], [0.0, 0.3]]),
+            2,
+            1.0,
+            0.01,
+            1e-2,
+        ),
+        (
+            1,
+            200_000,
+            tf.constant([[0.3, 0.2], [0.2, 0.22], [0.1, 0.25], [0.0, 0.3]]),
+            2,
+            2.0,
+            0.01,
+            1e-2,
+        ),
+        (2, 50_000, tf.constant([[0.0, 0.0]]), 2, 1.0, 0.01, 1e-2),
+    ],
+)
+def test_expected_hypervolume_improvement(
+    input_dim: int,
+    num_samples_per_point: int,
+    existing_observations: tf.Tensor,
+    obj_num: int,
+    variance_scale: float,
+    rtol: float,
+    atol: float,
+) -> None:
+    # Note: the test data number grows exponentially with num of obj
+    data_num_seg_per_dim = 10  # test data number per input dim
+    N = data_num_seg_per_dim ** input_dim
+    xs = tf.convert_to_tensor(
+        list(itertools.product(*[list(tf.linspace(-1, 1, data_num_seg_per_dim))] * input_dim))
+    )
+
+    xs = tf.cast(xs, dtype=existing_observations.dtype)
+    model = ModelStack(
+        *[(_mo_test_models[_](kernel_amplitude=variance_scale), 1) for _ in range(obj_num)]
+    )
+
+    mean, variance = model.predict(xs)
+
+    # [f_samples, B, L]
+    predict_samples = tfp.distributions.Normal(mean, tf.sqrt(variance)).sample(
+        num_samples_per_point
+    )
+    _pareto = Pareto(existing_observations)
+    ref_pt = get_reference_point(_pareto.front)
+    lb_points, ub_points = _pareto.hypercell_bounds(
+        tf.constant([-math.inf] * ref_pt.shape[-1]), ref_pt
+    )
+
+    # calc MC approx EHVI
+    splus_valid = tf.reduce_all(
+        tf.tile(ub_points[tf.newaxis, :, tf.newaxis, :], [num_samples_per_point, 1, N, 1])
+        > tf.expand_dims(predict_samples, axis=1),
+        axis=-1,
+    )  # num_cells x B
+    splus_idx = tf.expand_dims(tf.cast(splus_valid, dtype=ub_points.dtype), -1)
+    splus_lb = tf.tile(lb_points[tf.newaxis, :, tf.newaxis, :], [num_samples_per_point, 1, N, 1])
+    splus_lb = tf.maximum(splus_lb, tf.expand_dims(predict_samples, 1))
+    splus_ub = tf.tile(ub_points[tf.newaxis, :, tf.newaxis, :], [num_samples_per_point, 1, N, 1])
+    splus = tf.concat([splus_idx, splus_ub - splus_lb], axis=-1)
+
+    ehvi_approx = tf.transpose(
+        tf.reduce_sum(tf.reduce_prod(splus, axis=-1), axis=1, keepdims=True)
+    )  #
+    ehvi_approx = tf.reduce_mean(ehvi_approx, axis=-1)
+
+    ehvi = expected_hv_improvement(model, _pareto, ref_pt)(tf.expand_dims(xs, -2))
+
+    npt.assert_allclose(ehvi, ehvi_approx, rtol=rtol, atol=atol)
 
 
 @pytest.mark.parametrize("sample_size", [0, -2])
